feat: initial implementation of LoadedElement._build_model
